[refactor] (session-utils) move getSession to sessionUtils
